:bug:Fix : scheduler
- 스케줄러 2번씩 실행되던 문제해결

ref: #11
