+++ foodiehotspots/apps.py
@@ -6,7 +6,7 @@ class FoodiehotspotsConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'foodiehotspots'
 
-    def ready(self):
+    def ready(self): #앱 초기화 및 설정
         if settings.SCHEDULER_DEFAULT:
-            from config import tasks
+            from foodiehotspots import tasks
             tasks.start()
