+++ foodiehotspots/views.py
@@ -1,10 +1,11 @@
 from django.shortcuts import render
 from django.conf import settings
+from utils.get_data import processing_data
+
 logger = settings.CUSTOM_LOGGER
 
+processed_data = ''
 
 def test_scheduler():
-    logger.debug('This is a debug message')
-    logger.info('This is an info message')
-    logger.error('This is an error message')
-    print("test_scheduler!!!!!!!!")
+    logger.info("식당정보 얻어오기 start---!")
+    processed_data = processing_data()
